requirements.txt
Django~=3.1.3
pillow~=8.0.1

djangorestframework~=3.12.2
django-braces~=1.14.0